+++ pkg/chains/stellar/proto_helpers.go
@@ -1,6 +1,7 @@
 package stellar
 
 import (
+	"bytes"
 	"encoding/base64"
 	"encoding/hex"
 	"errors"
@@ -9,6 +10,7 @@ import (
 	stellarcap "github.com/smartcontractkit/chainlink-common/pkg/capabilities/v2/chain-capabilities/stellar"
 	"github.com/smartcontractkit/chainlink-common/pkg/capabilities/v2/chain-capabilities/stellar/scval"
 	"github.com/smartcontractkit/chainlink-common/pkg/types/chains/stellar"
+	"github.com/stellar/go-stellar-sdk/xdr"
 )
 
 // ConvertGetLedgerEntriesRequestToProto converts a domain GetLedgerEntriesRequest to its proto representation.
@@ -54,6 +56,26 @@ func ConvertGetLedgerEntriesRequestFromProto(p *GetLedgerEntriesRequest) (stella
 		keys[i] = base64.StdEncoding.EncodeToString(k)
 	}
 	return stellar.GetLedgerEntriesRequest{Keys: keys}, nil
+}
+
+func validateLedgerKeyXDR(b []byte) error {
+	if len(b) == 0 {
+		return fmt.Errorf("empty XDR")
+	}
+
+	var key xdr.LedgerKey
+	if err := key.UnmarshalBinary(b); err != nil {
+		return err
+	}
+
+	encoded, err := key.MarshalBinary()
+	if err != nil {
+		return err
+	}
+	if !bytes.Equal(encoded, b) {
+		return fmt.Errorf("trailing %d bytes", len(b)-len(encoded))
+	}
+	return nil
 }
 
 // ConvertLedgerEntryResultToProto converts a domain LedgerEntryResult to its proto representation.
